--- src/pages/Mainthema/Mainthema.scss
@@ -1,39 +1,0 @@
-.sortWrap {
-    padding-top: 150px;
-  padding-bottom: 100px;
-  display: flex;
-  position: absolute;
-  right: 0;
-  padding-right: 3vw;
-  justify-content: space-between;
-  width: 500px;
-  padding-top: 310px;
-  .label {
-    display: inline-block;
-    width: 160px;
-    height: 36px;
-    margin: 0 5px;
-    line-height: 34px;
-    text-align: center;
-    cursor: pointer;
-    font-weight: 500;
-    font-size: 1.3rem;
-    background-color: #fff;
-  }
-  .label:hover {
-    border-bottom: 1px solid #111;
-  }
-  input[type='radio'] {
-    -webkit-appearance: radio;
-  }
-  input[type='radio']:checked + .label {
-    border-color: #111;
-    background-color: #111;
-    color: #fff;
-  }
-  input {
-    position: absolute;
-    height: 0px;
-    overflow: hidden;
-  }
-}
