+++ src/pages/Maintheme/Maintheme.scss
@@ -2,3 +2,18 @@
     display: flex;
     width: 100vw;
 }
+
+.navUnderImg {
+    position: absolute;
+    left: 50%;
+    transform: translate(-50%);
+    width: 100%;
+    height: 200px;
+    top: 100px;
+  
+    img {
+      width: 100%;
+      height: 100%;
+    }
+  }
+
