+++ src/pages/Nav/Nav.scss
@@ -300,16 +300,3 @@
     }
   }
 }
-.navUnderImg {
-  position: absolute;
-  left: 50%;
-  transform: translate(-50%);
-  width: 100%;
-  height: 200px;
-  top: 100px;
-
-  img {
-    width: 100%;
-    height: 100%;
-  }
-}
